content(rules): update academic dishonesty phrasing
Update the phrasing of the Academic Dishonesty rule based on feedback discussed in Discord. The old phrasing made it sound like this was ok, but it was actually talking about what academic dishonesty and not that its ok to talk about in TPH. This was reworded to remove the source of confusion.